move getVersionString
