+++ Seeker/Main/MainActivity.cs
@@ -675,7 +675,7 @@ namespace Seeker
                     //var diag = builder.SetMessage(string.Format(SeekerState.ActiveActivityRef.GetString(Resource.String.about_body).TrimStart(' '), SeekerApplication.GetVersionString())).SetPositiveButton(Resource.String.close, OnCloseClick).Create();
                     var diag = builder.SetMessage(Resource.String.about_body).SetPositiveButton(Resource.String.close, OnCloseClick).Create();
                     diag.Show();
-                    var origString = string.Format(SeekerState.ActiveActivityRef.GetString(Resource.String.about_body), SeekerApplication.GetVersionString()); //this is a literal CDATA string.
+                    var origString = string.Format(SeekerState.ActiveActivityRef.GetString(Resource.String.about_body), CommonHelpers.GetVersionString()); //this is a literal CDATA string.
                     if (OperatingSystem.IsAndroidVersionAtLeast(24))
                     {
                         ((TextView)diag.FindViewById(Android.Resource.Id.Message)).TextFormatted = Android.Text.Html.FromHtml(origString, Android.Text.FromHtmlOptions.ModeLegacy); //this can be slow so do NOT do it in loops...
